Made building of .lib file limited to MinGW installations. Updated the arg count for MKMD functions in MKPerformSndMIDI_portaudio, cleaned up some types to keep the compiler happy on 64 bit systems

+++ MusicKit/Frameworks/MusicKit/_MKNameTable.h
@@ -36,7 +36,7 @@ Modification history:
     NSMapTable *types;
 }
 
-- (id) initWithCapacity: (unsigned) capacity;
+- (id) initWithCapacity: (NSUInteger) capacity;
 - (void) dealloc;
 
 @end
